+++ apps/passport-client/CHANGELOG.md
@@ -1,5 +1,32 @@
 # passport-client
 
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/eddsa-frog-pcd@0.1.0
+  - @pcd/eddsa-pcd@0.4.0
+  - @pcd/eddsa-ticket-pcd@0.4.0
+  - @pcd/email-pcd@0.4.0
+  - @pcd/emitter@0.3.0
+  - @pcd/ethereum-ownership-pcd@0.9.0
+  - @pcd/halo-nonce-pcd@0.5.0
+  - @pcd/passport-crypto@0.9.0
+  - @pcd/passport-interface@0.9.0
+  - @pcd/passport-ui@0.9.0
+  - @pcd/pcd-collection@0.9.0
+  - @pcd/pcd-types@0.9.0
+  - @pcd/rsa-image-pcd@0.3.0
+  - @pcd/rsa-pcd@0.4.0
+  - @pcd/rsa-ticket-pcd@0.4.0
+  - @pcd/semaphore-group-pcd@0.9.0
+  - @pcd/semaphore-identity-pcd@0.9.0
+  - @pcd/semaphore-signature-pcd@0.9.0
+  - @pcd/util@0.3.0
+  - @pcd/zk-eddsa-event-ticket-pcd@0.3.0
+  - @pcd/zk-eddsa-frog-pcd@0.1.0
+
 ## 0.0.24
 
 ### Patch Changes