+++ packages/passport-interface/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @pcd/passport-interface
 
+## 0.9.0
+
+### Minor Changes
+
+- esm+commonjs, compatibility, type declarations.
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/eddsa-frog-pcd@0.1.0
+  - @pcd/eddsa-pcd@0.4.0
+  - @pcd/eddsa-ticket-pcd@0.4.0
+  - @pcd/email-pcd@0.4.0
+  - @pcd/emitter@0.3.0
+  - @pcd/passport-crypto@0.9.0
+  - @pcd/pcd-collection@0.9.0
+  - @pcd/pcd-types@0.9.0
+  - @pcd/rsa-ticket-pcd@0.4.0
+  - @pcd/semaphore-group-pcd@0.9.0
+  - @pcd/semaphore-identity-pcd@0.9.0
+  - @pcd/semaphore-signature-pcd@0.9.0
+  - @pcd/util@0.3.0
+
 ## 0.8.0
 
 ### Minor Changes