+++ packages/rsa-image-pcd/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @pcd/rsa-image-pcd
 
+## 0.3.0
+
+### Minor Changes
+
+- esm+commonjs, compatibility, type declarations.
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/passport-ui@0.9.0
+  - @pcd/pcd-types@0.9.0
+  - @pcd/rsa-pcd@0.4.0
+
 ## 0.2.0
 
 ### Minor Changes